pyproject.toml
[project]
name = "upjobs"
version = "0.1.0"
description = "Upwork scraper"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "typer[all]",
  "playwright",
  "gspread",
  "google-auth",
  "google-auth-oauthlib",
  "python-dotenv",
  "pyyaml",
  "pandas",
  "numpy",
  "supabase",
  "gotrue",
  "jinja2",
  "httpx",
  "ipykernel>=6.30.1",
]

[project.scripts]
upjobs = "upjobs.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/upjobs"]

[tool.ruff]
line-length = 100

[tool.black]
line-length = 100
